requirements.txt
absl-py==0.12.0
alabaster==0.7.12
argh==0.26.2
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828549974/work
astor==0.8.1
astroid @ file:///C:/ci/astroid_1592481955828/work
astunparse==1.6.3
atomicwrites==1.4.0
attrs @ file:///tmp/build/80754af9/attrs_1598374659300/work
autopep8 @ file:///tmp/build/80754af9/autopep8_1596578164842/work
Babel==2.8.0
backcall==0.2.0
bcrypt @ file:///C:/ci/bcrypt_1597918112552/work
bleach==3.1.5
blinker==1.4
blis==0.7.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
catalogue==1.0.0
certifi==2020.6.20
cffi @ file:///C:/ci/cffi_1598352799320/work
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
colorama==0.4.3
commandlines==0.4.1
comtypes==1.1.8
cryptography @ file:///C:/ci/cryptography_1598873979408/work
cycler==0.10.0
cymem==2.0.5
Cython==0.29.14
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch @ file:///tmp/build/80754af9/diff-match-patch_1594828741838/work
docutils==0.16
editdistance==0.5.3
efficientnet==1.0.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
entrypoints==0.3
essential-generators==1.0
et-xmlfile==1.0.1
fer==20.1.1
filelock==3.0.12
flake8==3.8.3
Flask==1.1.2
flatbuffers==1.12
fonttools==4.21.1
future==0.18.2
gast==0.3.3
gensim==3.8.3
gevent==21.1.2
google-auth @ file:///tmp/build/80754af9/google-auth_1598987460909/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
greenlet==1.0.0
grpcio==1.32.0
gTTS==2.2.2
gunicorn==20.1.0
h5py==2.10.0
hsh==1.1.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio==2.9.0
imagesize==1.2.0
imantics==0.1.12
imgaug==0.4.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446525189/work
intervaltree @ file:///tmp/build/80754af9/intervaltree_1598376443606/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1598883894321/work
ipython-genutils==0.2.0
isort @ file:///C:/ci/isort_1597597762781/work
itsdangerous==1.1.0
jdcal==1.4.1
jedi @ file:///C:/ci/jedi_1592833832328/work
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema @ file:///C:/ci/jsonschema_1594363671836/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-core==4.6.3
jupyterlab==2.2.6
jupyterlab-server @ file:///tmp/build/80754af9/jupyterlab_server_1594164409481/work
Keras==2.1.5
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
keras-ocr==0.8.6
Keras-Preprocessing==1.1.2
keyring @ file:///C:/ci/keyring_1598866734005/work
kiwisolver==1.2.0
labelme2coco==0.1.2
lazy-object-proxy==1.4.3
llvmlite==0.36.0
lxml==4.6.3
Markdown @ file:///C:/ci/markdown_1597415185889/work
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib==3.3.1
mccabe==0.6.1
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mlxtend==0.17.3
mock==4.0.3
mtcnn==0.1.0
murmurhash==1.0.5
nbconvert @ file:///C:/ci/nbconvert_1594372737468/work
nbformat==5.0.7
networkx==2.5
nltk==3.5
notebook @ file:///C:/ci/notebook_1596837179121/work
numba==0.53.1
numpy==1.19.5
numpydoc @ file:///tmp/build/80754af9/numpydoc_1594166760263/work
oauthlib==3.1.0
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.42
openpyxl==3.0.6
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
paramiko @ file:///tmp/build/80754af9/paramiko_1598886428689/work
parso==0.7.0
pathtools==0.1.2
pexpect @ file:///C:/ci/pexpect_1594383465186/work
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
Pillow==7.2.0
pixellib==0.6.1
plac==1.1.3
playsound==1.2.2
pluggy==0.13.1
preshed==3.0.5
prometheus-client==0.8.0
prompt-toolkit==1.0.14
protobuf==3.13.0
psutil @ file:///C:/ci/psutil_1598352273540/work
pyasn1==0.4.8
pyasn1-modules==0.2.7
pyclipper==1.2.1
pycodestyle==2.6.0
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pydocstyle @ file:///tmp/build/80754af9/pydocstyle_1598885001695/work
pyflakes==2.2.0
Pygments==2.6.1
PyInquirer==1.0.3
PyJWT==1.7.1
pylint @ file:///C:/ci/pylint_1598605958085/work
PyMatting==1.1.2
PyNaCl @ file:///C:/ci/pynacl_1595009241355/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
pypiwin32==223
pyreadline==2.1
pyrsistent==0.16.0
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
python-jsonrpc-server @ file:///tmp/build/80754af9/python-jsonrpc-server_1594397536060/work
python-language-server @ file:///C:/ci/python-language-server_1594152769556/work
pyttsx3==2.90
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes @ file:///C:/ci/pywin32-ctypes_1594392691209/work
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
QDarkStyle==2.8.1
QtAwesome==0.7.2
qtconsole @ file:///tmp/build/80754af9/qtconsole_1598374667791/work
QtPy==1.9.0
regex==2020.11.13
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rope==0.17.0
rsa @ file:///tmp/build/80754af9/rsa_1596998415516/work
Rtree==0.9.4
sacremoses==0.0.43
scikit-image==0.18.1
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.0
Send2Trash==1.5.0
sentence-transformers==0.4.1.2
sentencepiece==0.1.95
Shapely==1.7.1
similaritymeasures==0.4.4
six==1.15.0
sklearn==0.0
smart-open==4.0.1
snowballstemmer==2.0.0
sortedcontainers==2.2.2
spacy==2.3.5
Sphinx @ file:///tmp/build/80754af9/sphinx_1597428793432/work
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder @ file:///C:/ci/spyder_1599052032150/work
spyder-kernels @ file:///C:/ci/spyder-kernels_1599051798778/work
srsly==1.0.5
tensorboard==2.4.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-hub==0.9.0
tensorflowjs==2.7.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-to-speech==1.5.2
tf-slim==1.1.0
thinc==7.4.5
threadpoolctl==2.1.0
tifffile==2021.3.5
tokenizers==0.9.4
toml @ file:///tmp/build/80754af9/toml_1592853716807/work
torch==1.8.0
torchvision==0.9.0
tornado==6.0.4
tqdm==4.54.0
traitlets==4.3.3
transformers==4.2.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
ujson==1.35
urllib3 @ file:///tmp/build/80754af9/urllib3_1597086586889/work
validators==0.18.2
wasabi==0.8.0
watchdog @ file:///C:/ci/watchdog_1593446515400/work
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==0.16.1
win-inet-pton==1.1.0
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.12.1
xgboost==1.2.0
xlrd==2.0.1
xmljson==0.2.1
yapf @ file:///tmp/build/80754af9/yapf_1593528177422/work
zipp==3.1.0
zope.event==4.5.0
zope.interface==5.3.0
